Move SD provider utils function to utils folder
