+++ examples/stable_diffusion/utils/ov.py
